foundry_agents/requirements.txt
# Microsoft Agent Framework with Foundry IQ (beta versions)
agent-framework>=1.0.0b260116
agent-framework-azure-ai-search>=1.0.0b260116

# Azure SDKs
azure-identity>=1.15.0
azure-search-documents>=11.6.0
azure-ai-projects>=1.0.0
openai>=1.40.0

# Web server
fastapi>=0.110.0
uvicorn>=0.27.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
pydantic>=2.0.0
